# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: |
    Fitness Agent Layer API — a single backend exposing fitness tools to both
    ChatGPT (Custom GPT Actions) and Claude (MCP server).
    
    Use userId to scope all state to a specific user. A simple string like
    "user_123" or an email is fine for single-user setups.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /state/{userId}:
    get:
      operationId: getState
      tags: [state]
      summary: Get full state for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Full fitness state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FitnessState"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: saveState
      tags: [state]
      summary: Save or update user fitness state
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveStateRequest"
      responses:
        "200":
          description: Updated fitness state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FitnessState"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    UserProfile:
      type: object
      properties:
        userId:
          type: string
        name:
          type: string
        age:
          type: integer
          nullable: true
        weightKg:
          type: number
          nullable: true
        heightCm:
          type: number
          nullable: true
        goal:
          type: string
          enum: [lose_weight, build_muscle, maintain, improve_endurance]
        allergies:
          type: array
          items:
            type: string
        preferences:
          type: array
          items:
            type: string
        budgetPerWeek:
          type: number
          nullable: true
        availableDays:
          type: array
          items:
            type: string
        sessionDurationMin:
          type: integer
          nullable: true
        equipment:
          type: array
          items:
            type: string
        injuries:
          type: array
          items:
            type: string
        mode:
          type: string
          enum: [auto, confirm]
      required: [userId, name, goal, allergies, preferences, availableDays, equipment, injuries, mode]
    Meal:
      type: object
      properties:
        name:
          type: string
        time:
          type: string
        calories:
          type: integer
        protein:
          type: number
        carbs:
          type: number
        fat:
          type: number
        ingredients:
          type: array
          items:
            type: string
      required: [name, calories, protein, carbs, fat, ingredients]
    Macros:
      type: object
      properties:
        proteinG:
          type: number
        carbsG:
          type: number
        fatG:
          type: number
      required: [proteinG, carbsG, fatG]
    DietPlan:
      type: object
      properties:
        meals:
          type: array
          items:
            $ref: "#/components/schemas/Meal"
        dailyCalories:
          type: integer
        macros:
          $ref: "#/components/schemas/Macros"
        notes:
          type: string
          nullable: true
      required: [meals, dailyCalories, macros]
    WorkoutSession:
      type: object
      properties:
        day:
          type: string
        name:
          type: string
        durationMin:
          type: integer
        exercises:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              sets:
                type: integer
              reps:
                type: integer
              restSec:
                type: integer
            required: [name]
      required: [day, name, durationMin, exercises]
    WorkoutPlan:
      type: object
      properties:
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutSession"
        notes:
          type: string
          nullable: true
      required: [sessions]
    ScheduleEvent:
      type: object
      properties:
        title:
          type: string
        date:
          type: string
          format: date
        time:
          type: string
        type:
          type: string
          enum: [workout, meal, check_in]
        durationMin:
          type: integer
      required: [title, date, type]
    Schedule:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/ScheduleEvent"
      required: [events]
    Achievement:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        earnedAt:
          type: string
          format: date-time
        xpBonus:
          type: integer
      required: [id, name, description, earnedAt, xpBonus]
    Reminder:
      type: object
      properties:
        id:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [missed_workout, missed_diet, weekly_report, plan_renewal]
        createdAt:
          type: string
          format: date-time
        read:
          type: boolean
      required: [id, message, type, createdAt, read]
    CompletionEvent:
      type: object
      properties:
        type:
          type: string
          enum: [workout, diet]
        completedAt:
          type: string
          format: date-time
        xpGained:
          type: integer
        notes:
          type: string
          nullable: true
      required: [type, completedAt, xpGained]
    Progress:
      type: object
      properties:
        xp:
          type: integer
        streak:
          type: integer
        level:
          type: integer
        history:
          type: array
          items:
            $ref: "#/components/schemas/CompletionEvent"
        achievements:
          type: array
          items:
            $ref: "#/components/schemas/Achievement"
        reminders:
          type: array
          items:
            $ref: "#/components/schemas/Reminder"
        lastLoggedAt:
          type: string
          format: date-time
          nullable: true
        xpToNextLevel:
          type: integer
      required: [xp, streak, level, history, achievements, reminders]
    FitnessState:
      type: object
      properties:
        profile:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/UserProfile"
        dietPlan:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/DietPlan"
        workoutPlan:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/WorkoutPlan"
        schedule:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/Schedule"
        progress:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/Progress"
      required: [profile, dietPlan, workoutPlan, schedule, progress]
    SaveStateRequest:
      type: object
      properties:
        profile:
          type: object
          properties:
            name:
              type: string
            age:
              type: integer
            weightKg:
              type: number
            heightCm:
              type: number
            goal:
              type: string
              enum: [lose_weight, build_muscle, maintain, improve_endurance]
            allergies:
              type: array
              items:
                type: string
            preferences:
              type: array
              items:
                type: string
            budgetPerWeek:
              type: number
            availableDays:
              type: array
              items:
                type: string
            sessionDurationMin:
              type: integer
            equipment:
              type: array
              items:
                type: string
            injuries:
              type: array
              items:
                type: string
            mode:
              type: string
              enum: [auto, confirm]
        dietPlan:
          $ref: "#/components/schemas/DietPlan"
        workoutPlan:
          $ref: "#/components/schemas/WorkoutPlan"
        schedule:
          $ref: "#/components/schemas/Schedule"
...